add JS-free variant of initial anim (breaks noRedir param)

+++ gen.sh
@@ -2,13 +2,18 @@
 set -euo pipefail
 IFS=$'\t\n'
 
-DOCS_PATH="${1:?need docs path}"
+./gen_blog.sh
 
-if [[ "$DOCS_PATH" == http* ]]; then
-  rm -rf tmp || true
-  git clone "$DOCS_PATH" tmp
-  DOCS_PATH="./tmp"
+if [ $# -ge 1 ]; then
+    DOCS_PATH="${1:?need docs path}"
+
+    if [[ "$DOCS_PATH" == http* ]]; then
+      rm -rf tmp || true
+      git clone "$DOCS_PATH" tmp
+      DOCS_PATH="./tmp"
+    fi
+
+    ./gen_blog.sh "$DOCS_PATH" ./thoughts_gen/ -m
 fi
 
-./gen_blog.sh
-./gen_blog.sh "$DOCS_PATH" ./thoughts_gen/ -m
+./nojs.sh
